feature: add sidepanel table for configuration
