+++ src/main/java/com/betternpchighlight/data/StyleDTO.java
@@ -1,0 +1,15 @@
+package com.betternpchighlight.data;
+
+public class StyleDTO {
+    public String tagStyle;
+    public int outlineColor; // ARGB int
+    public int fillColor;
+    public boolean raveOutline;
+    public boolean raveFill;
+    public int raveSpeed;
+    public String tileStyle;
+    // New fields
+    public double outlineWidth;
+    public boolean antiAliasing;
+    public int outlineFeather;
+}
