+++ src/main/java/com/betternpchighlight/util/IconSet.java
@@ -1,0 +1,17 @@
+package com.betternpchighlight.util;
+
+import javax.swing.*;
+
+public class IconSet {
+    public final ImageIcon on;
+    public final ImageIcon off;
+    public final ImageIcon onHover;
+    public final ImageIcon offHover;
+
+    public IconSet(ImageIcon on, ImageIcon off, ImageIcon onHover, ImageIcon offHover) {
+        this.on = on;
+        this.off = off;
+        this.onHover = onHover;
+        this.offHover = offHover;
+    }
+}
